Group org describe output and fold it into command.org.go
The handler and its tests move alongside the other org commands. Text
output now nests the OIDC values under an "OIDC Settings:" heading and the
trust-policy inputs under "AWS AssumeRole:", so the generic labels read in
context. Adds an e2e check that org describe returns an org_id.
